FIX: ad skipping and opt-out in ref app
There are a few things going on here:

We need to identify the duration of the trueX placeholder / blank video to skip over as part of the setup for interactive ads. Currently using the duration of the overall pod.

When completing an ad, the remainder of the ad break needs to be skipped, which corresponds to the ad break duration coming from Google, minus the aforementioned placeholder ad.

This fixes opt-in + complete ad, and opt-out. Auto-advance opt-out is broken because of player management issues (separate).